backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "google-genai>=1.19.0",
    "httpx>=0.28.1",
    "hyperbrowser>=0.45.1",
    "playwright>=1.52.0",
    "python-dotenv>=1.1.0",
]
